Created separate systemd services for uplink and bridge app
