+++ examples/rpi/config.toml
@@ -1,0 +1,3 @@
+[downloader]
+actions = ["update_firmware"]
+path = "/home/pi/download" # Where you want the update file to be downloaded
