+++ examples/rpi/device.json
@@ -1,0 +1,11 @@
+{
+  "project_id": "demo",
+  "broker": "cloud.bytebeam.io",
+  "port": 8883,
+  "device_id": "1245",
+  "authentication": {
+    "ca_certificate": "-----BEGIN CERTIFICATE-----\nMIIFrDCCA5SgAwIBAgICB+MwDQYJKoZIhvcNAQELBQAwdzEOMAwGA1UEBhMFSW5k\naWExETAPBgNVBAgTCEthcm5hdGFrMRIwEAYDVQQHEwlCYW5nYWxvcmUxFzAVBgNV\nBAkTDlN1YmJpYWggR2FyZGVuMQ8wDQYDVQQREwY1NjAwMTExFDASBgNVBAoTC0J5\ndGViZWFtLmlvMB4XDTIxMDkwMjExMDYyM1oXDTMxMDkwMjExMDYyM1owdzEOMAwG\nA1UEBhMFSW5kaWExETAPBgNVBAgTCEthcm5hdGFrMRIwEAYDVQQHEwlCYW5nYWxv\ncmUxFzAVBgNVBAkTDlN1YmJpYWggR2FyZGVuMQ8wDQYDVQQREwY1NjAwMTExFDAS\nBgNVBAoTC0J5dGViZWFtLmlvMIICIjANBgkqhkiG9w0BAQEFAAOCAg8AMIICCgKC\nAgEAr/bnOa/8AUGZmd/s+7rejuROgeLqqU9X15KKfKOBqcoMyXsSO65UEwpzadpw\nMl7GDCdHqFTymqdnAnbhgaT1PoIFhOG64y7UiNgiWmbh0XJj8G6oLrW9rQ1gug1Q\n/D7x2fUnza71aixiwEL+KsIFYIdDuzmoRD3rSer/bKOcGGs0WfB54KqIVVZ1DwsU\nk1wx5ExsKo7gAdXMAbdHRI2Szmn5MsZwGL6V0LfsKLE8ms2qlZe50oo2woLNN6XP\nRfRL4bwwkdsCqXWkkt4eUSNDq9hJsuINHdhO3GUieLsKLJGWJ0lq6si74t75rIKb\nvvsFEQ9mnAVS+iuUUsSjHPJIMnn/J64Nmgl/R/8FP5TUgUrHvHXKQkJ9h/a7+3tS\nlV2KMsFksXaFrGEByGIJ7yR4qu9hx5MXf8pf8EGEwOW/H3CdWcC2MvJ11PVpceUJ\neDVwE7B4gPM9Kx02RNwvUMH2FmYqkXX2DrrHQGQuq+6VRoN3rEdmGPqnONJEPeOw\nZzcGDVXKWZtd7UCbcZKdn0RYmVtI/OB5OW8IRoXFYgGB3IWP796dsXIwbJSqRb9m\nylICGOceQy3VR+8+BHkQLj5/ZKTe+AA3Ktk9UADvxRiWKGcejSA/LvyT8qzz0dqn\nGtcHYJuhJ/XpkHtB0PykB5WtxFjx3G/osbZfrNflcQZ9h1MCAwEAAaNCMEAwDgYD\nVR0PAQH/BAQDAgKEMA8GA1UdEwEB/wQFMAMBAf8wHQYDVR0OBBYEFKl/MTbLrZ0g\nurneOmAfBHO+LHz+MA0GCSqGSIb3DQEBCwUAA4ICAQAlus/uKic5sgo1d2hBJ0Ak\ns1XJsA2jz+OEdshQHmCCmzFir3IRSuVRmDBaBGlJDHCELqYxKn6dl/sKGwoqoAQ5\nOeR2sey3Nmdyw2k2JTDx58HnApZKAVir7BDxbIbbHmfhJk4ljeUBbertNXWbRHVr\ncs4XBNwXvX+noZjQzmXXK89YBsV2DCrGRAUeZ4hQEqV7XC0VKmlzEmfkr1nibDr5\nqwbI+7QWIAnkHggYi27lL2UTHpbsy9AnlrRMe73upiuLO7TvkwYC4TyDaoQ2ZRpG\nHY+mxXLdftoMv/ZvmyjOPYeTRQbfPqoRqcM6XOPXwSw9B6YddwmnkI7ohNOvAVfD\nwGptUc5OodgFQc3waRljX1q2lawZCTh58IUf32CRtOEL2RIz4VpUrNF/0E2vts1f\npO7V1vY2Qin998Nwqkxdsll0GLtEEE9hUyvk1F8U+fgjJ3Rjn4BxnCN4oCrdJOMa\nJCaysaHV7EEIMqrYP4jH6RzQzOXLd0m9NaL8A/Y9z2a96fwpZZU/fEEOH71t3Eo3\nV/CKlysiALMtsHfZDwHNpa6g0NQNGN5IRl/w1TS1izzjzgWhR6r8wX8OPLRzhNRz\n2HDbTXGYsem0ihC0B8uzujOhTHcBwsfxZUMpGjg8iycJlfpPDWBdw8qrGu8LeNux\na0cIevjvYAtVysoXInV0kg==\n-----END CERTIFICATE-----\n",
+    "device_certificate": "-----BEGIN CERTIFICATE-----\nMIIEajCCAlKgAwIBAgICB+MwDQYJKoZIhvcNAQELBQAwdzEOMAwGA1UEBhMFSW5k\naWExETAPBgNVBAgTCEthcm5hdGFrMRIwEAYDVQQHEwlCYW5nYWxvcmUxFzAVBgNV\nBAkTDlN1YmJpYWggR2FyZGVuMQ8wDQYDVQQREwY1NjAwMTExFDASBgNVBAoTC0J5\ndGViZWFtLmlvMB4XDTIzMDEwMzA1NDg1NloXDTMzMDEwMzA1NDg1NlowHjENMAsG\nA1UEChMEZGVtbzENMAsGA1UEAxMEMTI0NTCCASIwDQYJKoZIhvcNAQEBBQADggEP\nADCCAQoCggEBAPNjn0Fjwq0JPd9rd9tW8utj1VaOJ39Dz0Gt+0tUoXUsbwvWDvXQ\nUZFzgDZpKm+KgZBoDxfQCWzI4G1jFGkZTB+TBjU6Z9sL92T3F+6DU5vI7JdQsfeo\naKjoYJMLsAtOEO0l1Ajp/AmWAP1WLxAztNjU99ZyTCwlew1YkdypDtSnPG4AFhc7\nTFDAuSB8XTC1zDdXKkOuk5qVa9iEpQZ4uQU2oHc/sBSdfZzsy6kqG0T2y0348hTf\ncp+JlXT5D4iRHPk0Io4YO8ctlcjyt7vNzkyvLL+Evau+RxYg1tsrb22PEKu/tokV\nPNV5kC9xiFe/rPNek6HEYihRbcP2i8/x9h8CAwEAAaNZMFcwDgYDVR0PAQH/BAQD\nAgWgMBMGA1UdJQQMMAoGCCsGAQUFBwMCMB8GA1UdIwQYMBaAFKl/MTbLrZ0gurne\nOmAfBHO+LHz+MA8GA1UdEQQIMAaCBDEyNDUwDQYJKoZIhvcNAQELBQADggIBAAJM\nDYbDJ6qXq128ASKFol9O7NpKxYGCOjvXeCTkW+DaXH8gmUZKxcGGApmkagDaJ4vI\niCY4HlVlsoYGwKwMOJxaneHCtes89fhhHRNgRQjKHefki0NbyzjGRIgdvfQwGvXI\nxikkocJlHLLjh5yvFmME4W0HE6IGJbVDDLg6/eUmzgWbeQUPnpVpzWXO6SE6VtNV\nlWT43EktT546ioIwa0Lp5LAW6zmTStrWZj5W90AXrvzkdQIBhlhxUGfIRSlrW3vm\nYk98HidrefOrYL0I9ihyso1QLbfs7p7BlNRJK+ugHO4NIF29IQBKEUVDUUkcz5SA\n5aFQXEJBumR09rN9F8yOTpGNdewzJ/yHJII4SJRdOSiY5v0I58cLgeV1DRLJZLkK\n0XK8C/1R3REZhCZnxdt5GCteYicVzu3s4TUVa7DFUTvNXIxPua38+XUobhhsrGzu\n11HqieoMIwAkeu4zh2DtRmK1kv27ocfmRpgxHbSCompPP+Zoo2k8HsQzY8oyWCL6\nL6PhJtT3iE+akv8xdlNK0Kf1BnM2wk6cFAzScD2ziALrOq0WWj2kyYQBpEOFTwY+\nYP+SHugGJYKkOEjsrkf2XMQiUwAjk1eWyu18Aq1uBTk0iBEhe4e5xypQ/NlBAW1M\nzzImf04jsag/bZNGi4pMMVyG1tuX8f/y8s87+5xC\n-----END CERTIFICATE-----\n",
+    "device_private_key": "-----BEGIN RSA PRIVATE KEY-----\nMIIEogIBAAKCAQEA82OfQWPCrQk932t321by62PVVo4nf0PPQa37S1ShdSxvC9YO\n9dBRkXOANmkqb4qBkGgPF9AJbMjgbWMUaRlMH5MGNTpn2wv3ZPcX7oNTm8jsl1Cx\n96hoqOhgkwuwC04Q7SXUCOn8CZYA/VYvEDO02NT31nJMLCV7DViR3KkO1Kc8bgAW\nFztMUMC5IHxdMLXMN1cqQ66TmpVr2ISlBni5BTagdz+wFJ19nOzLqSobRPbLTfjy\nFN9yn4mVdPkPiJEc+TQijhg7xy2VyPK3u83OTK8sv4S9q75HFiDW2ytvbY8Qq7+2\niRU81XmQL3GIV7+s816TocRiKFFtw/aLz/H2HwIDAQABAoIBABbRQBzhPV77ftFz\nQclOXXez5Ew5B3rm4IdHyEdHp9tUTxTM/JoO7Za1PU/cjB3o3jyrhL62witwqZmX\nsFIqDPjoz6wtzY67kHXmtFen+C6ip4uIBScyop0zK9qdPePRbcTNxxMO3JIJfg5C\nBid+sXWigZmZHYsF+INHymTldH57AGHIAgMSjSvfADCo8j704QQQHvQK/oDm/QWO\nLH+yvCt2SlO7uMahAFabYQzh6LpXLy8FUmUcSEpDT14FYrfnkU4mZb9TULsua7ch\n/akxKSW5kPDu9VH60TT0ah4pNmuy+Mr3jMVXeC4upPC5EU+LgloFZTZWOWLaS/+q\nbVZkwAECgYEA+UII06jRwbqmtS0/0YUeZmCBSK7uCO59eLTrZBDUQsaMxRJOazU+\nAp98gPQDSuR7GOeky0v6sRBCc7f662FYCckNsLzATqix6XxfSY5riVI5Ey2SF0wK\nUx7pkEjrh8zuXGB9aXO8eSzOczPdjfnkTGDKfb4rMTFpd1Dq6LY84kMCgYEA+fjz\nGfg43VGYLfbbF1dLZiIfFr2rxNfdXp3MgGGsolHk8voGFMZMhgGd+Qex6d3R93l5\nbQRIMp7LZa7LcZl2Bkh+8umh7IPaDWar5bYZgVuufSHzuQBdJ077UHbHDBbnLwvp\nT+EriMG3plepOgH9TFeWyLvJ3pYWArNxwxD0JPUCgYA6Msbo1LQVLcWa6/PIsfJF\nKSdGl0GQ6E53jpafiKZm6QEZSorw0T+nUSG64j6aj/qF+CAsL5NbIFbDuiNmLhZm\nQoFl0Mx+ETlpMNb2z8AgEgith6hitCh8++zA+aW00Y6eFxFxO8q36FRT5bpHWh+K\nDdd5wJtJIyZi64vNxi81DQKBgE/hlVY0M3/ORlJR85uW4tplhwfOXujNjslgpexU\nykUo5IAdE/ppkCS2uZPOOLWilCmAYSRMXQIIKrl4zTylLMk/O4pGzN73F6R2rfH/\nalEjSjRkuCzjh+E65SbGfj9HxmMLbGiBbVcLiLOYidSmtBpEV0QYdp8x2SDN1Unf\nzhVNAoGAfAubNoPilTpm8HUd9pPnWIWDVlM1UkDbJeno5IL049DvfvRczmqqvPla\nL5q/a5SST0vn+fUhnRM6hSbTKKTT9QcsCYxI51ccbKuhJdKQWrscR73dfkcjLGFG\n9Jno2EwMW13nmS7LFM7Ai+uNMGHsvEVvLfawZvFPqNGeB3ynFrA=\n-----END RSA PRIVATE KEY-----\n"
+  }
+}
