+++ examples/rpi/systemd/bridge.sh
@@ -1,0 +1,2 @@
+# Run uplink
+python /mnt/download/bridge.py
