+++ examples/rpi/systemd/uplink.service
@@ -1,0 +1,13 @@
+[Unit]
+Description=Bytebeam Uplink service
+After=network.target auditd.service
+
+[Service]
+ExecStart=/mnt/download/systemd/uplink.sh
+KillMode=process
+Restart=on-failure
+Type=simple
+
+[Install]
+WantedBy=multi-user.target
+
