+++ examples/rpi/systemd/uplink.sh
@@ -1,0 +1,2 @@
+# Run uplink
+/mnt/download/uplink -c /mnt/download/config.toml -a /mnt/download/device.json -vvv
